Development

+++ features/src/main/java/com/kinley/features/flights/domain/Flight.kt
@@ -3,5 +3,5 @@ package com.kinley.features.flights.domain
 data class Flight(
     val flightName: String,
     val airlines: String,
-    val cost: Double
+    val cost: Int
 )
